refactor: #623 — relocate legacy/automation to top-level src/automation
CronScheduler + HeartbeatManager are independent runtime subsystems
(one runs cron jobs, the other periodically prompts agents). Both are
self-contained and used directly from app.ts; they belong at the top
level alongside other peer subsystems (workspace, sandbox, etc.).

- src/legacy/automation/cron-job{,.test}.ts → src/automation/
- src/legacy/automation/heartbeat{,.test}.ts → src/automation/
- src/legacy/automation/cron-config.test.ts folded into cron-job.test.ts
  (it was always testing CronScheduler, no separate cron-config.ts source)
- src/legacy/automation/index.ts deleted (dead barrel — app.ts and
  legacy/plugins/http imported file paths directly)
- Update consumer import paths (app.ts + legacy/plugins/http/*.ts)
- Drop stale path-banner header comments

src/legacy/automation/ directory is now gone — legacy/ shrinks to 7 entries.
