--- src/legacy/automation/index.ts
@@ -1,14 +1,0 @@
-export { CronScheduler } from "./cron-job.js";
-export type {
-  CronJob,
-  CronAction,
-  CronJobCallback,
-  CronJobInput,
-} from "./cron-job.js";
-
-export { HeartbeatManager } from "./heartbeat.js";
-export type {
-  HeartbeatConfig,
-  HeartbeatManagerOptions,
-  RunAgentLoop,
-} from "./heartbeat.js";
