+++ src/legacy/plugins/http/routes.ts
@@ -2,7 +2,7 @@
 
 import type { ServerResponse } from "node:http";
 
-import type { CronScheduler } from "../../automation/cron-job.js";
+import type { CronScheduler } from "../../../automation/cron-job.js";
 import type { SessionStoreManager } from "../../../agent/runners/pi/session-store.js";
 import type { HookRegistry } from "../../plugins/hooks.js";
 import type { AgentRuntime } from "../../../agent/runtime.js";
