+++ src/legacy/plugins/http/server.ts
@@ -4,7 +4,7 @@
 import http from "node:http";
 import path from "node:path";
 import { createLogger } from "../../../logging/logger.js";
-import type { CronScheduler } from "../../automation/cron-job.js";
+import type { CronScheduler } from "../../../automation/cron-job.js";
 import type { SessionStoreManager } from "../../../agent/runners/pi/session-store.js";
 import {
   applyCors,
